% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_metric}
\alias{calculate_metric}
\title{calculate_metric}
\usage{
calculate_metric(model, df, label, metric)
}
\arguments{
\item{model}{\code{object} a model object}

\item{df}{\code{data.frame} test data to score}

\item{label}{\code{numeric} target vector}

\item{metric}{\code{str} Machine learning metric to evaluate on. Choose between
\emph{roc_auc}, \emph{pr_auc}, \emph{F1}.}
}
\value{
\code{numeric} calculated ml metric
}
\description{
calculate_metric
}
