% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pp_score}
\alias{pp_score}
\title{pp_score}
\usage{
pp_score(df, x, y, cv_folds)
}
\arguments{
\item{df}{data.frame}

\item{x}{\code{str} name of features in \code{df}}

\item{y}{\code{str} target feature}

\item{cv_folds}{number of cross validations folds}
}
\value{
A list with the predictive power score for each feature in \code{x}
}
\description{
Calculates predictive power score for \code{x} predicts \code{y}. Where
\code{x} can be multiple features in \code{df}. This is a light (wip) implementation
which for now only supports binary classification and x as numeric features.
}
