% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_data}
\alias{sample_data}
\title{sample_data}
\usage{
sample_data(df, sample_size)
}
\arguments{
\item{df}{\code{data.frame} input data to sample from}

\item{sample_size}{\code{int} or \code{NULL} number of samples to take from \code{df}.}
}
\value{
\code{data.frame} with samples from \code{df} or \code{df} if \code{sample_size}
is \emph{NULL}.
}
\description{
Takes a random sample from \code{df} of size \code{sample_size}.
}
