% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{score}
\alias{score}
\title{Calculates predictive power score}
\usage{
score(df, x, y, cv_folds = 5L)
}
\arguments{
\item{df}{\code{data.frame} input data which contains \code{x} and \code{y}}

\item{x}{\code{str} name of feature}

\item{y}{\code{str} name of target}

\item{cv_folds}{\code{int} number of cross validations folds}
}
\value{
The predictive power score.
}
\description{
Calculates predictive power score for \code{x} predicts \code{y}. This is
a light wip implementation.
}
