% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pp_score}
\alias{pp_score}
\title{pp_score}
\usage{
pp_score(
  df,
  x,
  y,
  metric = "roc_auc",
  sample_size = NULL,
  cv_folds = 5,
  repeated_cv = 1
)
}
\arguments{
\item{df}{data.frame}

\item{x}{\code{str} name of features in \code{df}}

\item{y}{\code{str} target feature}

\item{metric}{\code{str} Machine learning metric to evaluate on. Choose between
\emph{roc_auc}, \emph{pr_auc}, \emph{F1}}

\item{sample_size}{\code{int} random sample taken from \code{df} to speed
up calculations. If \emph{NULL} all samples are used.}

\item{cv_folds}{number of cross validations folds}

\item{repeated_cv}{\code{int} number of repeated cross validations}
}
\value{
A list with the predictive power score for each feature in \code{x}
}
\description{
Calculates predictive power score for \code{x} predicts \code{y}. Where
\code{x} can be multiple features in \code{df}. This is a light (wip) implementation
which for now only supports binary classification and x as numeric features.
}
